+++ GestorCompeticiones/GestorCompeticionesHistoria2/src/logic/model/Inscripcion.java
@@ -12,7 +12,7 @@ public class Inscripcion {
 	public double tiempo;
 	public String medioPago;
 	public Date fechaPago;
-	public double cantidad;
+	public int cantidad;
 	public long atletaId;
 	public long competicionId;
 	public String nombreCompeticion;
